Import 4.9-1.15 release

+++ linux/drivers/soc/bcm/brcmstb/pm/xpt_dma.h
@@ -18,7 +18,7 @@
 
 struct dma_region {
 	dma_addr_t addr;
-	size_t len;
+	u64 len;
 	bool persistent;
 };
